pyproject.toml
[project]
name = "short"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "black>=26.1.0",
    "datetime>=6.0",
    "dnspython>=2.8.0",
    "fastapi>=0.133.1",
    "geoip2>=5.2.0",
    "httpx>=0.28.1",
    "jwt>=1.4.0",
    "logger>=1.4",
    "pathlib>=1.0.1",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "redis>=7.2.1",
    "rg>=0.0.76",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
    "validators>=0.35.0",
]

[dependency-groups]
dev = [
    "black>=26.1.0",
    "python-dotenv>=1.2.1",
]

